==> cards-scans analysis-scripts/Scripts- Segmentation Analysis/01-Descriptive Analysis 24-05.R <==
scans=read.csv("./scans.csv")
cards=read.csv("./cards.csv")
length(scans$card)
hist(scans$euros)
table(scans$category)
#getting in order of barcode
#same barcode has same products but vice-versa is not true
scans1<- scans[order(scans$barcode),]
date_unique<-unique(scans$date)
#843 unique dates
cat_unique<- unique(scans$category)
#24 unique categories
cards_unique<-unique(scans$card)
#124 unique cards or customers
bas_unique<-unique(scans$basket)
#67664 unique baskets taken by 124 customers 
#frequency vs Category
barplot(table(scans$category))
View(scans)
bas_unique
length(bas_unique)
table(scans$date)
summary(scans$date)
which.min(table(scans$date))
which.min(table(scans$date))
which.max(table(scans$date))
max(table(scans$date))
bar_unique
length(bar_unique)
barplot(table(scans$category),width = 0.2)
barplot(table(scans$category),width = 0.3)
barplot(table(scans$category),cex.names = 0.2)
barplot(table(scans$category),width = c(1,3),cex.names = 0.3)
barplot(table(scans$category),width = c(1,2),cex.names = 0.3)
barplot(table(scans$category),width = c(0.1,0.2),cex.names = 0.3)
hist(table(scans$date), breaks = 50)
barplot(table(scans$date),cex.names = 0.2,cex.axis = 0.3,xlab = "Date",ylab = "Number of Products")
table(scans$date)
scans_day<-scans
scans_day$day= weekdays(as.POSIXct(scans$date), abbreviate = F)
length(unique(scans_day$day))
unique(scans_day$day)
table(scans_day$day)
which.max(table(scans_day$day))
which.min(table(scans_day$day))
barplot(table(scans_day$day),cex.names = 0.5, cex.axis = 0.5,xlab = "Day",ylab = "Number of Products",col = c("Red","Blue"))
head(table(scans$category))
head(sort(table(scans$category)))
head(sort.default(table(scans$category)))
tail(sort(table(scans$category)))
#ggplot(tail(sort(table(scans$category))),cex.names = 0.5,cex.axis = 0.5,xlab = "Products",ylab = "Quantity",col = c("Red","Blue"))
b<-barplot(tail(sort(table(scans$category))),las=2,xlab = "Cat",ylab = "Quantity",cex.names=0.5,cex.axis=0.5)
text(b, y= tail(sort(table(scans$category)))/2-1, labels=as.character(tail(sort(table(scans$category)))))
c<-barplot(head(sort(table(scans$category))),las=2,xlab = "Cat",ylab = "Quantity",cex.names=0.5,cex.axis=0.5)
text(c, y= head(sort(table(scans$category)))/2-1, labels=as.character(head(sort(table(scans$category)))))
scans_card<-scans[order(scans$card),]
user_category<-data.frame(matrix(vector(), 336, 25,
                                 dimnames=list(c(), c("Card",as.character(cat_unique)))),
                          stringsAsFactors=FALSE)
user_category$Card<-cards$card
#colnames(user_category)
#match(user_category,"DRINKS")
#i="DRINKS"
#match(user_category,i)
#user_category[3,which(user_category==i,arr.ind = TRUE)]=1
#which(user_category=="DRINKS",arr.ind = TRUE)
#match.arg("DRINKS",user_category)
#match(user_category,"DRINKS")
user_cat<-data.frame(matrix(vector(), 548634, 2,
                             dimnames=list(c(), c("user","products"))),
                      stringsAsFactors=FALSE)
user_cat$user<-scans_card$card
user_cat$products<-scans_card$category
unique_user_cat<-unique(user_cat)
final<-as.data.frame.matrix(table(unique_user_cat))
yes <- 100 * colMeans(final[, 1:24] == "1")
dotchart(sort(yes), xlab = "Percent 'Users'",cex = 0.5 ,xlim = c(0, 100))
yes
max(cards$txn)
min(cards$txn)
summary(cards$gender)
mean(cards$age)
max(cards$units)
min(cards$units)
max(cards$euros)
min(cards$euros)
boxplot(cards$age,horizontal = TRUE, xlab= "Age")
mean(cards$days_per_txn)
#summary(scans$date)
as.Date(as.character("2014-09-30"), format="%Y-%m-%d")-as.Date(as.character(scans$date[1]), format="%Y-%m-%d")
#library(dplyr)
#table(scans$date)
#library(plyr)
table(scans_day$category,scans_day$day)
product_day<-as.data.frame.matrix(table(scans_day$category,scans_day$day))
rowSums(product_day)
barplot(rowSums(product_day),horiz = TRUE,xlim = c(0,200000),col = c("Red","Blue"),cex.names = 0.4,las=1, width = c(1,1),ylim = c(1,24))
scans1<-scans
library(dplyr)
dates <- as.Date(scans$date, "%Y-%m-%d")
years <- format(dates, "%Y")
tapply(scans$euros, years, sum)
b<-barplot(tapply(scans$euros, years, sum),cex.axis = 0.5,las=1,xlab = "Year",ylab = "Revenue in Euros")
text(b, y= tapply(scans$euros, years, sum)/2-1, labels=as.character(tapply(scans$euros, years, sum)),cex = 0.7)
dates <- as.Date(scans$date, "%Y-%m-%d")
months <- format(dates, "%Y-%m")
tapply(scans$euros, months, sum)
max(tapply(scans$euros, months, sum))
which.max(tapply(scans$euros, months, sum))
head(sort(tapply(scans$euros, months, sum)))
tail(sort(tapply(scans$euros, months, sum)))
pca_data<-as.data.frame.matrix(table(scans$card,scans$category))
pca_data_scaled<-scale(pca_data)
pca_data1<-prcomp(pca_data_scaled)
print(summary(pca_data1),digits=2)
library(ggplot2)
library(devtools)
install_github("vqv/ggbiplot")
library(ggbiplot)
ggbiplot(pca_data1,obs.scale = 1,var.scale = 1)
#plot(predict(pca_data1)[, 2:3], pch = 16,col = "grey")
#library("flexclust")
#projAxes(pca_data1, which = 2:3)
summary(cards$days_per_txn)
